pyproject.toml
[tool.poetry]
name = "online-beast"
version = "0.7.3"
description = ""
authors = ["Wytamma Wirth <wytamma.wirth@me.com>"]
readme = "README.md"

[tool.poetry.scripts]
online-beast = "online_beast.main:app"

[tool.poetry.dependencies]
python = ">=3.6.2,<4.0"
typer = {extras = ["all"], version = "^0.4.0"}
biopython = "^1.79"
lxml = "^4.7.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.1.0"
pre-commit = "^2.17.0"
pytest-cov = "^3.0.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
